+++ src/resources/views/back/clients/index.blade.php
@@ -145,7 +145,7 @@
 
                         <div class="col-md-6">
                             <div class="form-group">
-                                <label for="phone">Telefono <span class="text-info">(Opcional)</span></label>
+                                <label for="phone">Teléfono <span class="text-info">(Opcional)</span></label>
                                 <input type="text" name="phone" class="form-control">
                             </div>
                         </div>